Start dbus at boot
The first change is to start a dbus system session in the background
during boot. This is done by running the `dbus-daemon` command. We need
to create the directory that will contain the socket before we run the
command.

The second change is for the dbus user session. Instead of running sway
directly, we wrap it with the `dbus-run-session` command which runs it
inside of a dbus user session.