Add a lightweight seat management server and use it
A UNIX socket is created and given to sway in FD number 3. When sway
wants to open a device, it sends the path as a datagram to the socket.
poll detects that there is pending data on the socket and tells us about
it, we then process the message by opening the device and sending the FD
back to sway.